requirements-dev.txt
-r requirements.txt

pytest==3.5.1
pytest-cov==2.5.1
pytest-env==0.6.2
pytest-sugar==0.9.1
pytest-flask==0.10.0